+++ widget/view.controller.js
@@ -1,9 +1,15 @@
+/* Copyright start
+  MIT License
+  Copyright (c) 2024 Fortinet Inc
+  Copyright end */
+ 
+ 'use strict';
 (function () {
     angular
         .module('cybersponse')
-        .controller('customPicklistMessage111Ctrl', customPicklistMessage111Ctrl);
-        customPicklistMessage111Ctrl.$inject = ['$scope', '$state', '$interval', 'Modules', 'config', 'websocketService', '$rootScope', '_'];
-    function customPicklistMessage111Ctrl($scope, $state, $interval, Modules, config, websocketService, $rootScope, _ ) {
+        .controller('customPicklistMessage112Ctrl', customPicklistMessage112Ctrl);
+        customPicklistMessage112Ctrl.$inject = ['$scope', '$state', '$interval', 'Modules', 'config', 'websocketService', '$rootScope', '_'];
+    function customPicklistMessage112Ctrl($scope, $state, $interval, Modules, config, websocketService, $rootScope, _ ) {
         $scope.config = config;
         $scope.title = '';
         $scope.timeinterval = 0;
@@ -48,12 +54,8 @@
         );
         
         function widgetWSSubscribe() {
-            if (widgetsubscription) {
-                websocketService.unsubscribe(widgetsubscription);
-            }
             websocketService.subscribe($state.params.module+'/'+$state.params.id, function (result) {
                 var changedAttribute;
-                widgetsubscription = result;
                 if (angular.isDefined(result.changeData)) {
                     if (result.changeData.includes(config.picklistItem)) {
                         changedAttribute = config.picklistItem;
@@ -62,8 +64,9 @@
                 if (changedAttribute) {
                    checkPicklistCondition();
                 }
-            }
-            );
+            }).then(function (data){
+					widgetsubscription=data;
+			});
         }
         widgetWSSubscribe();
         checkPicklistCondition();