+++ widget/release_notes.md
@@ -1,5 +1,5 @@
 ## Release Notes
 
-Following enhancements have been made to the custom picklist message widget in version `v1.1.1`:
+Following enhancements have been made to the custom picklist message widget in version `v1.1.2`:
 
-- Optimized the API calls this widget triggered resulting in improved performance
+- Fixed Minor Bug